requirements.txt
Django==5.0.6
psycopg2_binary==2.9.2
djangorestframework==3.15.2
django-filter==21.1
drf-nested-routers==0.94.1
djangorestframework-simplejwt==5.3.1
